Created scanner for config files
Successfully scanned 7 config files.
Updates to the scanner will be needed during the development of the parser.

Known issues:
No support for float (now it interprets a floats as two integer and a dot).
To simplify the multi-line comment support the file was preprocessed and
these comments were removed. This results in incorrect line number refferences.